requirements.txt
python-dotenv==1.0.1
playwright==1.49.1
beautifulsoup4==4.12.3
markdownify==0.14.1
tiktoken==0.8.0
langchain-text-splitters==0.3.4
langchain-cerebras==0.4.0
langchain-openai==0.2.14
pymilvus==2.5.1
torch==2.5.1
langchain==0.3.13
langgraph==0.2.60
thefuzz==0.22.1
pydantic==2.10.1
chainlit==1.3.2
tokenizers==0.19.1
sentence-transformers==3.3.1
fastembed==0.5.0
milvus-model==0.2.11
transformers==4.44.2
einops==0.8.0
joblib==1.4.2
FlagEmbedding==1.3.3
fastapi[all]==0.115.6
